djangoproj/requirements.txt
django
django-ninja
psycopg2-binary
whitenoise
python-dotenv
django-environ
python-jose
boto3
